Tests/LibWeb/Text/input/css-has-invalidation/baseline-cache-hits.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script src="_helpers.js"></script>
<style>
    /* Many sibling anchors with the same :has() selector. Each sibling will be
       tested during a single style pass. The HasResultCache should absorb repeat
       matches against the same anchor. */
    .anchor:has(.match) { color: red; }
    .anchor:has(.match) + .anchor { color: blue; }
</style>
<div id="container">
    <div class="anchor"><span>x</span></div>
    <div class="anchor"><span>x</span></div>
    <div class="anchor"><span>x</span></div>
    <div class="anchor"><span>x</span></div>
    <div class="anchor"><span class="match" id="target">x</span></div>
</div>
<script>
    test(() => {
        getComputedStyle(document.getElementById("container")).color;
        internals.resetStyleInvalidationCounters();

        // Removing the .match class forces re-match on all anchors in the next pass.
        document.getElementById("target").classList.remove("match");
        getComputedStyle(document.getElementById("container")).color;
        printCounters("after remove .match");
    });
</script>
