Tests/LibWeb/Text/expected/css-has-invalidation/baseline-cache-hits.txt

[after remove .match]
  hasAncestorWalkInvocations: 1
  hasAncestorWalkVisits: 5
  hasMatchInvocations: 2
  hasResultCacheHits: 0
  hasResultCacheMisses: 2
  styleInvalidations: 1
